wip: infrastructure for backend plugin, without backand calls yet...
